Update Spelling / Employer
My employer does not define me as a computer scientist.

They are merely a vehicle to my retirement that stands in between my otherwise lifelong collaborative computer science work for the greater humanity and the united state's unique relationship with debt and worker exploitation.

Additionally my name was spelled wrong and that specific employer was very outdated. Please merge.

Signed-off-by: Kris Nóva <kris@nivenly.com>